MAINT: Add BarReader base class for both minute and session readers
