+++ zipline/data/minute_bars.py
@@ -11,7 +11,6 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-from abc import ABCMeta, abstractmethod, abstractproperty
 import json
 import os
 from os.path import join
@@ -23,7 +22,6 @@ from bcolz import ctable
 from intervaltree import IntervalTree
 import numpy as np
 import pandas as pd
-from six import with_metaclass
 from toolz import keymap, valmap
 
 from zipline.data._minute_bar_internal import (
@@ -33,6 +31,8 @@ from zipline.data._minute_bar_internal import (
 )
 
 from zipline.gens.sim_engine import NANOS_IN_MINUTE
+
+from zipline.data.bar_reader import BarReader
 from zipline.utils.calendars import get_calendar
 from zipline.utils.cli import maybe_show_progress
 from zipline.utils.memoize import lazyval
@@ -53,106 +53,10 @@ class BcolzMinuteWriterColumnMismatch(Exception):
     pass
 
 
-class MinuteBarReader(with_metaclass(ABCMeta)):
-
-    _data_frequency = 'minute'
-
+class MinuteBarReader(BarReader):
     @property
     def data_frequency(self):
-        return self._data_frequency
-
-    @abstractproperty
-    def last_available_dt(self):
-        """
-        Returns
-        -------
-        dt : pd.Timestamp
-            The last minute for which the reader can provide data.
-        """
-        pass
-
-    @abstractproperty
-    def first_trading_day(self):
-        """
-        Returns
-        -------
-        dt : pd.Timestamp
-            The first trading day (session) for which the reader can provide
-            data.
-        """
-        pass
-
-    @abstractmethod
-    def get_value(self, sid, dt, field):
-        """
-        Retrieve the value at the given coordinates.
-
-        Parameters
-        ----------
-        sid : int
-            The asset identifier.
-        dt : pd.Timestamp
-            The minute label for the desired data point.
-        field : string
-            The OHLVC name for the desired data point.
-
-        Returns
-        -------
-        value : float|int
-            The value at the given coordinates, ``float`` for OHLC, ``int``
-            for 'volume'.
-        """
-        pass
-
-    @abstractmethod
-    def get_last_traded_dt(self, asset, dt):
-        """
-        Get the latest minute on or before ``dt`` in which ``asset`` traded.
-
-        If there are no trades on or before ``dt`` returns ``pd.NaT``
-
-        Parameters
-        ----------
-        asset : zipline.asset.Asset
-            The asset for which to get the last traded minute.
-        dt : pd.Timestamp
-            The minute at which to start searching for the last traded minute.
-
-        Returns
-        -------
-        last_traded : pd.Timestamp
-            The minute of the last trade for the given asset, using the input
-            dt as a vantage point.
-        """
-        pass
-
-    @abstractmethod
-    def load_raw_arrays(self, fields, start_dt, end_dt, sids):
-        """
-        Retrieve the arrays of pricing data for the given coordinates of
-        ``fields`` (OHLCV), minute range [``start_dt``, ``end_dt``] and sids.
-
-        Parameters
-        ----------
-        fields : iterable of str
-            The OHLCV fields ('open', 'high', 'low', 'close', 'volume') for
-            which to read data.
-        start_dt : pd.Timestamp
-            The first minute of the date range for which to read data.
-        end_dt : pd.Timestamp
-            The last minute of the date range for which to read data.
-        sids : iterable of int
-            The sid identifiers for which to retrieve data.
-
-        Returns
-        -------
-        raw_arrays : list of ndarray
-            A list where each item corresponds with the fields in the order
-            the fields are given.
-            Each item is a 2D array with a shape of (minutes_in_range, sids)
-            The OHLC arrays are floats; the 'volume' array is ints.
-        """
-        pass
+        return "minute"
 
 
 def _calc_minute_index(market_opens, minutes_per_day):
